# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Texas Childcare Advisors Compliance Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type LocationFacilityType = typeof LocationFacilityType[keyof typeof LocationFacilityType];


export const LocationFacilityType = {
  child_care_center: 'child_care_center',
  licensed_home: 'licensed_home',
  school_age: 'school_age',
} as const;

export interface Location {
  id: number;
  name: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zip?: string | null;
  facilityType: LocationFacilityType;
  createdAt: string;
  updatedAt: string;
}

export type LocationInputFacilityType = typeof LocationInputFacilityType[keyof typeof LocationInputFacilityType];


export const LocationInputFacilityType = {
  child_care_center: 'child_care_center',
  licensed_home: 'licensed_home',
  school_age: 'school_age',
} as const;

export interface LocationInput {
  /** @minLength 1 */
  name: string;
  address?: string;
  city?: string;
  state?: string;
  zip?: string;
  facilityType?: LocationInputFacilityType;
}

export type LocationUpdateFacilityType = typeof LocationUpdateFacilityType[keyof typeof LocationUpdateFacilityType];


export const LocationUpdateFacilityType = {
  child_care_center: 'child_care_center',
  licensed_home: 'licensed_home',
  school_age: 'school_age',
} as const;

export interface LocationUpdate {
  /** @minLength 1 */
  name?: string;
  address?: string;
  city?: string;
  state?: string;
  zip?: string;
  facilityType?: LocationUpdateFacilityType;
}

export type StaffMemberStatus = typeof StaffMemberStatus[keyof typeof StaffMemberStatus];


export const StaffMemberStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface StaffMember {
  id: number;
  locationId: number;
  firstName: string;
  lastName: string;
  /** @nullable */
  email?: string | null;
  role: string;
  status: StaffMemberStatus;
  /** @nullable */
  hireDate?: string | null;
  /** @nullable */
  yearsExperience?: number | null;
  annualTrainingHours: number;
  preserviceHours: number;
  createdAt: string;
  updatedAt: string;
}

export type StaffMemberInputStatus = typeof StaffMemberInputStatus[keyof typeof StaffMemberInputStatus];


export const StaffMemberInputStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface StaffMemberInput {
  locationId: number;
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  email?: string;
  /** @minLength 1 */
  role: string;
  status?: StaffMemberInputStatus;
  hireDate?: string;
  yearsExperience?: number;
  annualTrainingHours?: number;
  preserviceHours?: number;
}

export type StaffMemberUpdateStatus = typeof StaffMemberUpdateStatus[keyof typeof StaffMemberUpdateStatus];


export const StaffMemberUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface StaffMemberUpdate {
  locationId?: number;
  /** @minLength 1 */
  firstName?: string;
  /** @minLength 1 */
  lastName?: string;
  email?: string;
  /** @minLength 1 */
  role?: string;
  status?: StaffMemberUpdateStatus;
  hireDate?: string;
  yearsExperience?: number;
  annualTrainingHours?: number;
  preserviceHours?: number;
}

export interface CertificationType {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  isDefault: boolean;
  /** @nullable */
  validityMonths?: number | null;
  createdAt: string;
}

export interface CertificationTypeInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  validityMonths?: number;
}

export type CertificationWithStatusStatus = typeof CertificationWithStatusStatus[keyof typeof CertificationWithStatusStatus];


export const CertificationWithStatusStatus = {
  valid: 'valid',
  expiring: 'expiring',
  expired: 'expired',
  no_expiry: 'no_expiry',
} as const;

export interface CertificationWithStatus {
  id: number;
  staffId: number;
  certificationTypeId: number;
  certificationTypeName: string;
  staffFirstName: string;
  staffLastName: string;
  locationId: number;
  locationName: string;
  /** @nullable */
  issuedDate?: string | null;
  /** @nullable */
  expirationDate?: string | null;
  /** @nullable */
  notes?: string | null;
  status: CertificationWithStatusStatus;
  /** @nullable */
  daysUntilExpiration?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CertificationInput {
  staffId: number;
  certificationTypeId: number;
  issuedDate?: string;
  expirationDate?: string;
  notes?: string;
}

export interface CertificationUpdate {
  certificationTypeId?: number;
  issuedDate?: string;
  expirationDate?: string;
  notes?: string;
}

export interface LocationComplianceSummary {
  locationId: number;
  locationName: string;
  totalStaff: number;
  compliantStaff: number;
  expiringSoonCount: number;
  expiredCount: number;
}

export interface FreeTierUsage {
  staffCount: number;
  staffLimit: number;
  locationCount: number;
  locationLimit: number;
}

export interface DashboardSummary {
  totalStaff: number;
  compliantStaff: number;
  expiringSoonStaff: number;
  expiredCertifications: number;
  totalCertifications: number;
  locationBreakdown: LocationComplianceSummary[];
  freeTierUsage: FreeTierUsage;
}

export interface RisingStarRequirement {
  description: string;
  met: boolean;
  /** @nullable */
  detail?: string | null;
}

export interface RisingStarCategory {
  name: string;
  score: number;
  maxScore: number;
  requirements: RisingStarRequirement[];
}

export interface RisingStarScore {
  locationId: number;
  locationName: string;
  currentLevel: number;
  /** @nullable */
  nextLevel?: number | null;
  overallScore: number;
  categories: RisingStarCategory[];
  recommendations: string[];
}

export type ListStaffParams = {
locationId?: number;
status?: ListStaffStatus;
};

export type ListStaffStatus = typeof ListStaffStatus[keyof typeof ListStaffStatus];


export const ListStaffStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export type ListCertificationsParams = {
staffId?: number;
locationId?: number;
status?: ListCertificationsStatus;
};

export type ListCertificationsStatus = typeof ListCertificationsStatus[keyof typeof ListCertificationsStatus];


export const ListCertificationsStatus = {
  valid: 'valid',
  expiring: 'expiring',
  expired: 'expired',
} as const;

export type GetDashboardParams = {
locationId?: number;
};

export type ListExpiringCertificationsParams = {
locationId?: number;
daysAhead?: number;
};

export type GetRisingStarScoreParams = {
locationId: number;
};

export type ExportCsvParams = {
locationId?: number;
};