# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Texas Childcare Advisors Compliance Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List all locations for the authenticated user
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      summary: Create a new location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Created location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Free tier limit reached (3 locations max)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /locations/{id}:
    get:
      operationId: getLocation
      tags: [locations]
      summary: Get a location by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLocation
      tags: [locations]
      summary: Update a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdate"
      responses:
        "200":
          description: Updated location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLocation
      tags: [locations]
      summary: Delete a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get compliance dashboard summary
      parameters:
        - name: locationId
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
...
  /rising-star:
    get:
      operationId: getRisingStarScore
      tags: [dashboard]
      summary: Get Rising Star compliance score for a location
      parameters:
        - name: locationId
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Rising Star score and breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RisingStarScore"
        "404":
          description: Location not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /export/csv:
    get:
      operationId: exportCsv
      tags: [dashboard]
      summary: Export staff certification data as CSV
      parameters:
        - name: locationId
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: CSV file content
          content:
            text/csv:
              schema:
                type: string
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Location:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zip:
          type: ["string", "null"]
        facilityType:
          type: string
          enum: [child_care_center, licensed_home, school_age]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - facilityType
        - createdAt
        - updatedAt
    LocationInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        facilityType:
          type: string
          enum: [child_care_center, licensed_home, school_age]
      required:
        - name
    LocationUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        facilityType:
          type: string
          enum: [child_care_center, licensed_home, school_age]
...
    DashboardSummary:
      type: object
      properties:
        totalStaff:
          type: number
        compliantStaff:
          type: number
        expiringSoonStaff:
          type: number
        expiredCertifications:
          type: number
        totalCertifications:
          type: number
        locationBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/LocationComplianceSummary"
        freeTierUsage:
          $ref: "#/components/schemas/FreeTierUsage"
      required:
        - totalStaff
        - compliantStaff
        - expiringSoonStaff
        - expiredCertifications
        - totalCertifications
        - locationBreakdown
        - freeTierUsage
    LocationComplianceSummary:
      type: object
      properties:
        locationId:
          type: number
        locationName:
          type: string
        totalStaff:
          type: number
        compliantStaff:
          type: number
        expiringSoonCount:
          type: number
        expiredCount:
          type: number
      required:
        - locationId
        - locationName
        - totalStaff
        - compliantStaff
        - expiringSoonCount
        - expiredCount
    FreeTierUsage:
      type: object
      properties:
        staffCount:
          type: number
        staffLimit:
          type: number
        locationCount:
          type: number
        locationLimit:
          type: number
      required:
        - staffCount
        - staffLimit
        - locationCount
        - locationLimit
    RisingStarScore:
      type: object
      properties:
        locationId:
          type: number
        locationName:
          type: string
        currentLevel:
          type: number
        nextLevel:
          type: ["number", "null"]
        overallScore:
          type: number
        categories:
          type: array
          items:
            $ref: "#/components/schemas/RisingStarCategory"
        recommendations:
          type: array
          items:
            type: string
      required:
        - locationId
        - locationName
        - currentLevel
        - overallScore
        - categories
        - recommendations
    RisingStarCategory:
      type: object
      properties:
        name:
          type: string
        score:
          type: number
        maxScore:
          type: number
        requirements:
          type: array
          items:
            $ref: "#/components/schemas/RisingStarRequirement"
      required:
        - name
        - score
        - maxScore
        - requirements
    RisingStarRequirement:
      type: object
      properties:
        description:
          type: string
        met:
          type: boolean
        detail:
          type: ["string", "null"]
      required:
        - description
        - met