# lib/api-zod/src/generated/types/location.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Texas Childcare Advisors Compliance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { LocationFacilityType } from './locationFacilityType';

export interface Location {
  id: number;
  name: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zip?: string | null;
  facilityType: LocationFacilityType;
  createdAt: string;
  updatedAt: string;
}