# lib/api-zod/src/generated/types/locationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Texas Childcare Advisors Compliance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { LocationInputFacilityType } from './locationInputFacilityType';

export interface LocationInput {
  /** @minLength 1 */
  name: string;
  address?: string;
  city?: string;
  state?: string;
  zip?: string;
  facilityType?: LocationInputFacilityType;
}