# lib/api-zod/src/generated/types/locationUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Texas Childcare Advisors Compliance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { LocationUpdateFacilityType } from './locationUpdateFacilityType';

export interface LocationUpdate {
  /** @minLength 1 */
  name?: string;
  address?: string;
  city?: string;
  state?: string;
  zip?: string;
  facilityType?: LocationUpdateFacilityType;
}